Fix: new points system, 1/3 of old points

+++ models/drawWindow.cpp
@@ -233,7 +233,7 @@ void DrawWindow::drawRect(int startX, int startY, int width, int heigth,
 }
 
 void DrawWindow::drawStats(int startX, int startY, int width, int heigth,
-                           long *points, Character character,
+                           float *points, Character character,
                            pEnemyList enemyList, int powerUp) {
   int powerUp_y = 52;
   int powerUp_x = 23;
@@ -248,7 +248,7 @@ void DrawWindow::drawStats(int startX, int startY, int width, int heigth,
 
   init_pair(3, COLOR_YELLOW, -1);  // FUNZIONI PER USARE I COLORI
   attron(COLOR_PAIR(3));
-  mvprintw(startX - 2, startY + 12, "%lu", *points);
+  mvprintw(startX - 2, startY + 12, "%.0f", *points);
   if (character.getNumberLife() == 3)
     mvprintw(startX - 2, startY + 38, "[C] [C] [C]");
   if (character.getNumberLife() == 2)
@@ -259,7 +259,7 @@ void DrawWindow::drawStats(int startX, int startY, int width, int heigth,
 
   init_pair(3, COLOR_YELLOW, -1);  // FUNZIONI PER USARE I COLORI
   attron(COLOR_PAIR(3));
-  mvprintw(startX - 2, startY + 12, "%lu", *points);
+  mvprintw(startX - 2, startY + 12, "%.0f", *points);
   if (powerUp == 4)
     mvprintw(powerUp_x, powerUp_y + 10, "[X] [X] [X] [X]");
   if (powerUp == 3)
